refactor migration

+++ api/database/migrations/2025_04_02_024152_create_users_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('username', 100)->nullable();
             $table->string('email', 255);
             $table->string('phone', 40)->nullable();
-            $table->unsignedInteger('customer_id')->nullable();
+            $table->unsignedBigInteger('customer_id')->nullable();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password', 255);
             $table->string('remember_token', 100)->nullable();
